refactor(drive): 收敛回收站空状态文案
